pyproject.toml
[tool.poetry]
name = "mush_app"
version = "0.1.0"
description = ""
authors = ["Graham-Broughton <broughton.graham@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pandas = "^2.1.3"
numpy = "^1.26.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
tensorflow = "^2.10.0"
tensorflow-addons = "^0.22.0"
wandb = "^0.16.0"

[tool.poetry.group.deploy.dependencies]
uvicorn = "^0.24.0.post1"
fastapi = "^0.104.1"
python-multipart = "^0.0.6"
twilio = "^8.10.1"
pillow = "^10.1.0"

[tool.poetry.group.train.dependencies]
tqdm = "^4.66.1"
prefect = "^2.14.4"
scikit-learn = "^1.3.2"
mlflow = "^2.8.0"
opencv-python = "^4.8.1.78"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
